Update 0.0.1 md
Changed PPF/s to PPF. 

Note--Although PPF is used interchangeably with umol/s, the standard unit of measurement is the latter: umol/s. I recommend the latter unit expression (umol/s) because it follows the other sensor unit expressions. PPFD of course factors in area in terms of m2, so: umol/m2/s.